refactor: extract shared EncryptedDek conversions to own target
Move InternalEncryptedDek.toPublic()/EncryptedDek.toInternal() (and their ProtobufFormat helpers) out of RankIndexBlobService.kt into a new service/v1alpha/EncryptedDekConversions.kt (:encrypted_dek_conversions) so PoolAssignmentJobService and RawImpressionUploadModelLineService can reuse them. The two conversion functions become public as they now cross bazel module boundaries.
